feat: Dynamically load cards from deck
