+++ app/src/main/res/layout/activity_game_screen.xml
@@ -13,5 +13,4 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:name="at.aau.serg.fragments.CardsFragment" />
-
 </androidx.constraintlayout.widget.ConstraintLayout>
