+++ app/src/main/res/layout/fragment_card.xml
@@ -7,6 +7,7 @@
     >
 
     <ImageView
+        android:id="@+id/cardImageView"
         android:layout_width="200dp"
         android:layout_height="wrap_content"
         android:scaleType="fitCenter"
